Allow any RecipeType (in theory)
