+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/auto_crafters/SlimefunItemRecipe.java
@@ -9,29 +9,28 @@ import javax.annotation.Nonnull;
 
 import org.bukkit.inventory.ItemStack;
 
-import io.github.thebusybiscuit.slimefun4.implementation.items.multiblocks.EnhancedCraftingTable;
 import io.github.thebusybiscuit.slimefun4.implementation.tasks.AsyncRecipeChoiceTask;
 import io.github.thebusybiscuit.slimefun4.utils.ChestMenuUtils;
 import io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils;
 import me.mrCookieSlime.CSCoreLibPlugin.general.Inventory.ChestMenu;
+import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 
 /**
  * This {@link AbstractRecipe} implementation stands for a {@link SlimefunItem} which
- * is crafted using an {@link EnhancedCraftingTable}.
+ * is crafted using any {@link RecipeType}.
  * 
  * @author TheBusyBiscuit
  * 
- * @see EnhancedCraftingTable
- * @see EnhancedAutoCrafter
+ * @see SlimefunAutoCrafter
  *
  */
-class EnhancedRecipe extends AbstractRecipe {
+class SlimefunItemRecipe extends AbstractRecipe {
 
     private final int[] slots = { 11, 12, 13, 20, 21, 22, 29, 30, 31 };
     private final SlimefunItem item;
 
-    EnhancedRecipe(@Nonnull SlimefunItem item) {
+    SlimefunItemRecipe(@Nonnull SlimefunItem item) {
         super(getInputs(item), item.getRecipeOutput());
         this.item = item;
     }
